+++ lib/syoboi_calendar/song.rb
@@ -1,40 +1,21 @@
 module SyoboiCalendar
   class Song
-    KEY_ARRANGERS = "編曲"
-    KEY_COMPOSERS = "作曲"
-    KEY_LYRIC_WRITERS = "作詞"
-
     # @return [String]
     attr_reader :name
 
+    # @return [Array<SyoboiCalendar::Personality>]
+    attr_reader :personalities
+
+    # @return [String]
+    attr_reader :role
+
     # @param name [String]
-    # @param source [Hash]
-    def initialize(name:, source:)
+    # @param personalities [Array<SyoboiCalendar::Personality>]
+    # @param role [String]
+    def initialize(name:, personalities:, role:)
       @name = name
-      @source = source
-    end
-
-    # @return [Array<String>]
-    def arrangers
-      source[KEY_ARRANGERS]
-    end
-
-    # @return [Array<String>]
-    def composers
-      source[KEY_COMPOSERS]
-    end
-
-    # @return [Array<String>]
-    def lyric_writers
-      source[KEY_LYRIC_WRITERS]
-    end
-
-    private
-
-    # @private
-    # @return [Hash]
-    def source
-      @source
+      @personalities = personalities
+      @role = role
     end
   end
 end